requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.0+cpu
azure-functions==1.20.0
azure-storage-blob==12.20.0
azure-identity==1.15.0
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
opencv-python-headless
fastapi
uvicorn